backend/requirements.txt
pandas
numpy
fastapi
uvicorn
python-multipart
python-multipart
pdfplumber
pytesseract
pdf2image
python-dotenv
httpx
pdfminer.six